Implement OverloadSet solution for Clang.
Unlike the original Clang <17 specific solution this one works one abstraction level higher where primary expression can be used in dcltype. That way we can use structured binadings as the actual test in an unevaluated context which is where clang 17 breaks the older implementaiton.

The new implementation still does not work for GCC. That compiler still after years has a parser bug with the necessary code.