Cleanup (#117)
Cleanup done with DWYU.
Not fully working yet as it cannot be done with Clang yet.
DWYU would need to understand Google Style BUILD file comments that prevent actions. Or at least that needs to be doable with buildozer.. but with the latter alone DWYU would need to do all changes in one go.